Add missing database calls for 'highestUnique'

+++ app/src/androidTest/java/com/qrpokemon/qrpokemon/LeaderboardActivityTest.java
@@ -44,7 +44,7 @@ public class LeaderboardActivityTest {
     public void checkBackButton() {
         onView(withId(R.id.leaderboard_back)).perform(click());
 
-        // NOTE: Expresso stays open which prevents getting the DESTROYED state
+        // NOTE: Espresso stays open which prevents getting the DESTROYED state
         // Therefore, check that the state was changed by the button
         assertNotEquals(Lifecycle.State.RESUMED, activityRule.getScenario().getState());
     }
@@ -140,14 +140,14 @@ public class LeaderboardActivityTest {
             @Override
             public void updateRank(Context context, LeaderboardList list) {
                 ((LeaderboardActivity) context).setPersonalRank(mockItem.getRank(), mockItem.getUsername(),
-                        (int) mockItem.getHighestScore(), (int) mockItem.getQrQuantity(), (int) mockItem.getTotalScore());
+                        mockItem.getHighestScore(), mockItem.getQrQuantity(), mockItem.getTotalScore());
 
                 // Checking here ensures the data has updated
                 onView(withId(R.id.tv_leaderboard_player_rank)).check(matches(withText(mockItem.getRank())));
                 onView(withId(R.id.tv_leaderboard_player_username)).check(matches(withText(mockItem.getUsername())));
-                onView(withId(R.id.tv_leaderboard_player_unique)).check(matches(withText((int) mockItem.getHighestScore())));
-                onView(withId(R.id.tv_leaderboard_player_qrcount)).check(matches(withText((int) mockItem.getQrQuantity())));
-                onView(withId(R.id.tv_leaderboard_player_score)).check(matches(withText((int) mockItem.getTotalScore())));
+                onView(withId(R.id.tv_leaderboard_player_unique)).check(matches(withText(mockItem.getHighestScore())));
+                onView(withId(R.id.tv_leaderboard_player_qrcount)).check(matches(withText(mockItem.getQrQuantity())));
+                onView(withId(R.id.tv_leaderboard_player_score)).check(matches(withText(mockItem.getTotalScore())));
             }
         }
